framework: merge inputmanager impl
+fix npe in gazecursorcontroller by passing inputmgr to ctor
